Dodao da organizator ne može brisati tuđe natjecanje.

+++ frontend/src/stranice/natjecanja.jsx
@@ -36,6 +36,16 @@ export default function Natjecanja() {
     }, [competitions]);
 
    const obrisiNatjecanje = async () => {
+        if (!odabranoNatjecanje) return;
+        const response = await fetch(`${IP}/natjecanja/${odabranoNatjecanje._id}`, {credentials: "include"});
+        const data = await response.json();
+        console.log(data);
+
+        if (data.organizatorId != korisnik._id) {
+            alert('Nemate dopuštenje brisati ovo natjecanje.');
+            return;
+        }
+
         try {
             const response = await fetch (`${IP}/natjecanja/${odabranoNatjecanje._id}`, {
                 method: "DELETE",
